テーマテンプレートの実装を助ける目的ではないメソッドを `$theme` から分離して隠蔽した。
